% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_datasets.R
\docType{data}
\name{BLUEs.pheno}
\alias{BLUEs.pheno}
\title{Best Linear Unbiased Estimates of phenotype}
\format{
An object of class \code{data.frame} with 50 rows and 2 columns.
}
\usage{
BLUEs.pheno
}
\description{
Best Linear Unbiased Estimates of phenotype
}
\keyword{datasets}
