% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_datasets.R
\docType{data}
\name{bx}
\alias{bx}
\title{Rcpp internal function Backward from forward-backward algorithm}
\format{
An object of class \code{function} of length 1.
}
\usage{
bx
}
\description{
Rcpp internal function Backward from forward-backward algorithm
}
\keyword{datasets}
