% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_functions.R
\name{findQTLpeaks}
\alias{findQTLpeaks}
\title{Extract QTL peak positions from the results of a QTL scan}
\usage{
findQTLpeaks(LOD_data)
}
\arguments{
\item{LOD_data}{Output of \code{\link{QTLscan}} function}
}
\description{
Function to find QTL peaks from output of a QTL scan
}
