% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_datasets.R
\docType{data}
\name{fx}
\alias{fx}
\title{Rcpp internal function Forward from forward-backward algorithm}
\format{
An object of class \code{function} of length 1.
}
\usage{
fx
}
\description{
Rcpp internal function Forward from forward-backward algorithm
}
\keyword{datasets}
