% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_datasets.R
\docType{data}
\name{GIC_4x}
\alias{GIC_4x}
\title{A list of GIC estimates (4x)}
\format{
An object of class \code{list} of length 2.
}
\usage{
GIC_4x
}
\description{
A list of GIC estimates (4x)
}
\keyword{datasets}
