% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_datasets.R
\docType{data}
\name{hexa.list}
\alias{hexa.list}
\title{A list of hexaploid bivalent pairing configurations}
\format{
An object of class \code{list} of length 15.
}
\usage{
hexa.list
}
\description{
A list of hexaploid bivalent pairing configurations
}
\keyword{datasets}
