% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_functions.R
\name{list.depth}
\alias{list.depth}
\title{Find depth of a list}
\usage{
list.depth(obj, objdepth = 0)
}
\arguments{
\item{obj}{An input object, may be a (nested) list}

\item{objdepth}{Counter to record how deep the recursion has gone}
}
\description{
Recursive function checking list depth to determine subsequent subsetting rules
}
