% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_functions.R
\name{mapseq}
\alias{mapseq}
\title{Generate a sequence of map positions for splining function}
\usage{
mapseq(cMvect, gap)
}
\arguments{
\item{cMvect}{Vector of map positions}

\item{gap}{Gap size in cM}
}
\description{
Function to return a sequence of map positions at steps of size gap from given input
}
