% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_datasets.R
\docType{data}
\name{mr.ls}
\alias{mr.ls}
\title{Example output of meiosis report function}
\format{
An object of class \code{list} of length 2.
}
\usage{
mr.ls
}
\description{
Example output of meiosis report function
}
\keyword{datasets}
