% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_functions.R
\name{NettletonDoerge}
\alias{NettletonDoerge}
\title{Nettleton and Doerge 2000}
\usage{
NettletonDoerge(N = 100, alpha = 0.05, gamma = 0.05)
}
\arguments{
\item{N}{number of permutations}

\item{alpha}{The threshold level, ie. (1 - alpha) quantile of sorted LOD scores defines threshold}

\item{gamma}{The confidence interval specifier, usually 0.05}
}
\description{
Nettleton and Doerge 2000
}
