% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_functions.R
\name{Nstates.fun}
\alias{Nstates.fun}
\title{Error handling on ploidy and ploidy2, and determine the number of genotype classes for general ploidy level and pairing behaviour}
\usage{
Nstates.fun(biv_dec, pl, pl2)
}
\arguments{
\item{biv_dec}{Either TRUE for bivalents only or FALSE to also allow quadrivalents and double reduction}

\item{pl}{ploidy level of parent 1, assumed to be even and greater than 4}

\item{pl2}{ploidy level of parent 2, assumed to be even and greater than 2}
}
\description{
Error handling on ploidy and ploidy2, and determine the number of genotype classes for general ploidy level and pairing behaviour
}
