% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_datasets.R
\docType{data}
\name{phased_maplist.4x}
\alias{phased_maplist.4x}
\title{A list of phased maps (4x)}
\format{
An object of class \code{list} of length 2.
}
\usage{
phased_maplist.4x
}
\description{
A list of phased maps (4x)
}
\keyword{datasets}
