% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_datasets.R
\docType{data}
\name{Phenotypes_4x}
\alias{Phenotypes_4x}
\title{Phenotypes for example tetraploid}
\format{
An object of class \code{data.frame} with 150 rows and 3 columns.
}
\usage{
Phenotypes_4x
}
\description{
Phenotypes for example tetraploid
}
\keyword{datasets}
