% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyqtlR-package.R
\docType{package}
\name{polyqtlR-package}
\alias{polyqtlR}
\alias{polyqtlR-package}
\title{QTL analysis in polyploid species using posterior genotype probabilities}
\description{
Quantitative trait loci (QTL) analysis and exploration of meiotic patterns in autopolyploid bi-parental F1 populations. For all ploidy levels, identity-by-descent (IBD) probabilities can be estimated. Significance thresholds, exploring QTL allele effects and visualising results are provided. For more background and to reference the package see \doi{10.1093/bioinformatics/btab574}.
}
\author{
\strong{Maintainer}: Peter Bourke \email{pbourkey@gmail.com}

Other contributors:
\itemize{
  \item Christine Hackett [contributor]
  \item Chris Maliepaard [contributor]
  \item Geert van Geest [contributor]
  \item Roeland Voorrips [contributor]
  \item Johan Willemsen [contributor]
}

}
\keyword{internal}
