% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyqtlR_functions.R
\docType{package}
\name{polyqtlR}
\alias{polyqtlR}
\title{QTL analysis in polyploid species using identity-by-descent probabilities}
\description{
R package to perform QTL analysis using marker data from polyploid species.
}
