% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_functions.R
\name{probgeno_df_to_array}
\alias{probgeno_df_to_array}
\title{Convert a progeno_df data.frame to a 3d array}
\usage{
probgeno_df_to_array(probgeno_df, ploidy)
}
\arguments{
\item{probgeno_df}{A data frame as read from the scores file produced by function
\code{saveMarkerModels} of R package \code{fitPoly}, or alternatively, a data frame containing at least the following columns:
\itemize{
\item{SampleName : }{
Name of the sample (individual)
}
\item{MarkerName : }{
Name of the marker
}
\item{P0}{
Probabilities of dosage score '0'
}
\item{P1, P2... etc. : }{
Probabilities of dosage score '1' etc. (up to max offspring dosage, e.g. P4 for tetraploid population)
}
}}

\item{ploidy}{Ploidy of the F1 population (can be 2, 3, 4 or 6)}
}
\description{
Convert a progeno_df data.frame to a 3d array
}
