% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_datasets.R
\docType{data}
\name{qtl_LODs.4x}
\alias{qtl_LODs.4x}
\title{A list of QTL results (4x)}
\format{
An object of class \code{list} of length 4.
}
\usage{
qtl_LODs.4x
}
\description{
A list of QTL results (4x)
}
\keyword{datasets}
