% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_functions.R
\name{quadTM}
\alias{quadTM}
\title{Quadrivalent transition matrix function}
\usage{
quadTM(r)
}
\arguments{
\item{r}{recombination frequency}
}
\description{
Quadrivalent transition matrix function
}
