% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_datasets.R
\docType{data}
\name{Rec_Data_4x}
\alias{Rec_Data_4x}
\title{A list of recombination count data (4x)}
\format{
An object of class \code{list} of length 2.
}
\usage{
Rec_Data_4x
}
\description{
A list of recombination count data (4x)
}
\keyword{datasets}
