% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_datasets.R
\docType{data}
\name{rem.hex}
\alias{rem.hex}
\title{Redundant genotype classes in hexavalent transition matrix (6x)}
\format{
An object of class \code{integer} of length 166.
}
\usage{
rem.hex
}
\description{
Redundant genotype classes in hexavalent transition matrix (6x)
}
\keyword{datasets}
