% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_datasets.R
\docType{data}
\name{rem.quad}
\alias{rem.quad}
\title{Redundant genotype classes in quadrivalent transition matrix (4x)}
\format{
An object of class \code{integer} of length 6.
}
\usage{
rem.quad
}
\description{
Redundant genotype classes in quadrivalent transition matrix (4x)
}
\keyword{datasets}
