% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_datasets.R
\docType{data}
\name{segList_2x}
\alias{segList_2x}
\title{A list of all possible bi-allelic QTL segregation types (2x)}
\format{
An object of class \code{list} of length 8.
}
\usage{
segList_2x
}
\description{
A list of all possible bi-allelic QTL segregation types (2x)
}
\keyword{datasets}
