% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_datasets.R
\docType{data}
\name{segList_3x}
\alias{segList_3x}
\title{A list of all possible bi-allelic QTL segregation types (3x)}
\format{
An object of class \code{list} of length 27.
}
\usage{
segList_3x
}
\description{
A list of all possible bi-allelic QTL segregation types (3x)
}
\keyword{datasets}
