% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_datasets.R
\docType{data}
\name{segList_3x_24}
\alias{segList_3x_24}
\title{Expected segregation for all markers types of a triploid cross (2 x 4)}
\format{
An object of class \code{list} of length 27.
}
\usage{
segList_3x_24
}
\description{
Expected segregation for all markers types of a triploid cross (2 x 4)
}
\keyword{datasets}
