% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_datasets.R
\docType{data}
\name{segList_4x}
\alias{segList_4x}
\title{Expected segregation for all markers types of a tetraploid cross}
\format{
An object of class \code{list} of length 224.
}
\usage{
segList_4x
}
\description{
Expected segregation for all markers types of a tetraploid cross
}
\keyword{datasets}
