% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_datasets.R
\docType{data}
\name{segList_6x}
\alias{segList_6x}
\title{A list of all possible bi-allelic QTL segregation types (6x)}
\format{
An object of class \code{list} of length 3735.
}
\usage{
segList_6x
}
\description{
A list of all possible bi-allelic QTL segregation types (6x)
}
\keyword{datasets}
