% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_functions.R
\name{state_fun}
\alias{state_fun}
\title{Function to return the list of possible pairing states, given parental ploidies and meiotic pairing model}
\usage{
state_fun(ploidy, ploidy2, bivalent_decoding)
}
\arguments{
\item{ploidy}{Ploidy of parent 1}

\item{ploidy2}{Ploidy of parent 2}

\item{bivalent_decoding}{Logical, if \code{FALSE} then multivalent pairing assumed}
}
\description{
Function to return the list of possible pairing states, given parental ploidies and meiotic pairing model
}
