% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_functions.R
\name{test_IBD_list}
\alias{test_IBD_list}
\title{Error and warning handling for IBD_list as estimated by \code{estimate_IBD}}
\usage{
test_IBD_list(IBD_list)
}
\arguments{
\item{IBD_list}{List of IBD probabilities}
}
\description{
Error and warning handling for IBD_list as estimated by \code{estimate_IBD}
}
