% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_functions.R
\name{TM.biv.2}
\alias{TM.biv.2}
\title{Bivalent-pairing transition matrix function, diploid}
\usage{
TM.biv.2(m1, r_vect)
}
\arguments{
\item{m1}{marker}

\item{r_vect}{Vector of adjacent recombination frequencies}
}
\description{
Bivalent-pairing transition matrix function, diploid
}
