% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_functions.R
\name{TM.biv.4}
\alias{TM.biv.4}
\title{Bivalent-pairing transition matrix function, tetraploid}
\usage{
TM.biv.4(m1, r_vect)
}
\arguments{
\item{m1}{marker}

\item{r_vect}{Vector of adjacent recombiantion frequencies}
}
\description{
Bivalent-pairing transition matrix function, tetraploid
}
