% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_functions.R
\name{TM.hex}
\alias{TM.hex}
\title{Hexavalent-pairing transition matrix function, hexaploid}
\usage{
TM.hex(m1, r_vect, rem_hex = rem.hex)
}
\arguments{
\item{m1}{marker}

\item{r_vect}{Vector of adjacent recombiantion frequencies}

\item{rem_hex}{Index of genotype classes to remove, work-around from overly-general transition matrix with redundant classes}
}
\description{
Hexavalent-pairing transition matrix function, hexaploid
}
