% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_functions.R
\name{TM.quad}
\alias{TM.quad}
\title{Quadrivalent-pairing transition matrix function, tetraploid}
\usage{
TM.quad(m1, r_vect, rem_quad = rem.quad)
}
\arguments{
\item{m1}{marker}

\item{r_vect}{Vector of adjacent recombination frequencies}

\item{rem_quad}{Index of genotype classes to remove, work-around from overly-general transition matrix with redundant classes}
}
\description{
Quadrivalent-pairing transition matrix function, tetraploid
}
