% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_functions.R
\name{TMfun.2x_B}
\alias{TMfun.2x_B}
\title{Diploid bi-parental transition matrix}
\usage{
TMfun.2x_B(m1, r_vect)
}
\arguments{
\item{m1}{marker}

\item{r_vect}{Vector of adjacent recombiantion frequencies}
}
\description{
Diploid bi-parental transition matrix
}
