% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_functions.R
\name{TMfun.3x_BB}
\alias{TMfun.3x_BB}
\title{Triploid bi-parental transition matrix, bivalent-bivalent pairing}
\usage{
TMfun.3x_BB(...)
}
\arguments{
\item{\dots}{Arguments passed to parental TM functions}
}
\description{
Triploid bi-parental transition matrix, bivalent-bivalent pairing
}
