% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_functions.R
\name{TMfun.4x_BB}
\alias{TMfun.4x_BB}
\title{Tetraploid bi-parental transition matrix, bivalent-bivalent pairing}
\usage{
TMfun.4x_BB(...)
}
\arguments{
\item{\dots}{Arguments passed to parental TM functions}
}
\description{
Tetraploid bi-parental transition matrix, bivalent-bivalent pairing
}
