% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_functions.R
\name{TMfun.4x_QQ}
\alias{TMfun.4x_QQ}
\title{Tetraploid bi-parental transition matrix, quadrivalent-quadrivalent pairing}
\usage{
TMfun.4x_QQ(...)
}
\arguments{
\item{\dots}{Arguments passed to parental TM functions}
}
\description{
Tetraploid bi-parental transition matrix, quadrivalent-quadrivalent pairing
}
