% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_functions.R
\name{TMfun.6x_BH}
\alias{TMfun.6x_BH}
\title{Hexaploid bi-parental transition matrix, bivalent-hexavalent pairing}
\usage{
TMfun.6x_BH(...)
}
\arguments{
\item{\dots}{Arguments passed to parental TM functions}
}
\description{
Hexaploid bi-parental transition matrix, bivalent-hexavalent pairing
}
