% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_functions.R
\name{TMfun.6x_HB}
\alias{TMfun.6x_HB}
\title{Hexaploid bi-parental transition matrix, hexavalent-bivalent pairing}
\usage{
TMfun.6x_HB(...)
}
\arguments{
\item{\dots}{Arguments passed to parental TM functions}
}
\description{
Hexaploid bi-parental transition matrix, hexavalent-bivalent pairing
}
