% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_functions.R
\name{weighted.var}
\alias{weighted.var}
\title{Calculate the weighted variance}
\usage{
weighted.var(x, w, na.rm = FALSE)
}
\arguments{
\item{x}{Vector of interest}

\item{w}{Vector of weights}

\item{na.rm}{Should missing values be removed? By default, \code{FALSE}}
}
\description{
Generalisation of the variance to include weights
}
