% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_datasets.R
\docType{data}
\name{BLUEs.pheno}
\alias{BLUEs.pheno}
\title{A data-frame of best linear unbiased predicted (BLUE) phenotypes (4x)}
\format{
An object of class \code{data.frame} with 50 rows and 2 columns.
}
\usage{
BLUEs.pheno
}
\description{
A data-frame of best linear unbiased predicted (BLUE) phenotypes (4x)
}
\keyword{datasets}
