% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_datasets.R
\docType{data}
\name{GIC_4x}
\alias{GIC_4x}
\title{Genotypic Information Coefficient for example tetraploid}
\format{
An object of class \code{list} of length 2.
}
\usage{
GIC_4x
}
\description{
Genotypic Information Coefficient for example tetraploid
}
\keyword{datasets}
