% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_datasets.R
\docType{data}
\name{IBD_4x}
\alias{IBD_4x}
\title{A list of identity-by-descent probabilities (4x)}
\format{
An object of class \code{list} of length 2.
}
\usage{
IBD_4x
}
\description{
A list of identity-by-descent probabilities (4x)
}
\keyword{datasets}
