% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_datasets.R
\docType{data}
\name{qtl_LODs.4x}
\alias{qtl_LODs.4x}
\title{QTL output for example tetraploid}
\format{
An object of class \code{list} of length 6.
}
\usage{
qtl_LODs.4x
}
\description{
QTL output for example tetraploid
}
\keyword{datasets}
