% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_datasets.R
\docType{data}
\name{segList_2x}
\alias{segList_2x}
\title{Expected segregation for all markers types of a diploid cross}
\format{
An object of class \code{list} of length 8.
}
\usage{
segList_2x
}
\description{
Expected segregation for all markers types of a diploid cross
}
\keyword{datasets}
