% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_datasets.R
\docType{data}
\name{segList_3x}
\alias{segList_3x}
\title{Expected segregation for all markers types of a triploid cross (4 x 2)}
\format{
An object of class \code{list} of length 27.
}
\usage{
segList_3x
}
\description{
Expected segregation for all markers types of a triploid cross (4 x 2)
}
\keyword{datasets}
