% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_datasets.R
\docType{data}
\name{segList_4x}
\alias{segList_4x}
\title{A list of all possible bi-allelic QTL segregation types (4x)}
\format{
An object of class \code{list} of length 224.
}
\usage{
segList_4x
}
\description{
A list of all possible bi-allelic QTL segregation types (4x)
}
\keyword{datasets}
