% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_datasets.R
\docType{data}
\name{segList_6x}
\alias{segList_6x}
\title{Expected segregation for all markers types of a hexaploid cross}
\format{
An object of class \code{list} of length 3735.
}
\usage{
segList_6x
}
\description{
Expected segregation for all markers types of a hexaploid cross
}
\keyword{datasets}
