% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_datasets.R
\docType{data}
\name{SNP_dosages.4x}
\alias{SNP_dosages.4x}
\title{SNP marker dosage data for example tetraploid}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 186 rows and 52 columns.
}
\usage{
SNP_dosages.4x
}
\description{
SNP marker dosage data for example tetraploid
}
\keyword{datasets}
